Email_Automation-Py-main/requirements.txt
Flask>=3.0.0
pymongo>=4.6.0
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
requests>=2.31.0
flask-login>=0.6.3
Werkzeug>=3.0.0
python-dotenv>=1.0.0
distro>=1.9.0
sympy>=1.13.0
ollama